+++ src/configuration.rs
@@ -45,6 +45,8 @@ pub enum SubCommand {
 pub struct Configuration {
     // Source files.
     pub source_files: Vec<PathBuf>,
+    // Object files to be linked.
+    pub object_files: Vec<PathBuf>,
     // Runs memory sanitizer to detect memory leak and invalid memory reference at early time.
     // Requires shared library ,/sanitizer/libfixsanitizer.so.
     pub sanitize_memory: bool,
@@ -151,6 +153,7 @@ impl Configuration {
         Ok(Configuration {
             subcommand,
             source_files: vec![],
+            object_files: vec![],
             sanitize_memory: false,
             fix_opt_level: FixOptimizationLevel::Default, // Fix's optimization level.
             linked_libraries: vec![],